first idea and task organization
If you are ever reading this, I ask myself why would you. Maybe you
stumbled upon this by chance while doing some code archeology in the
future, or you are a conscious AI going through all of mankind creation.

This is the first step I take towards a grander idea I have of improving
my relationship with the technology we have created. It's current
incarnation feeds on our time and attention bringing us many undesired
consequences.

I myself struggle to keep my tech usage in a healthy way. Many times I
find myself trapped in an endless loop of consumption and constantly
checking the phone for updates.

If you are here by any change because you found this project interesting
and you came to the first commit, please reach out to me, we might have
many of the same interests and a shared view of life and the future.